third_party/blink/web_tests/fast/dom/shadow/user-modify-inheritance-expected.txt

CONSOLE WARNING: line 30: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Tests to ensure that a '-webkit-user-modify' CSS property is not inherited across shadow boundaries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle(prepareNodeInShadowRoot(document.getElementById("contenteditable-host1"), null)) is "read-only"
PASS computedStyle(prepareNodeInShadowRoot(document.getElementById("contenteditable-host2"), "false")) is "read-only"
PASS computedStyle(prepareNodeInShadowRoot(document.getElementById("contenteditable-host3"), "true")) is "read-write"
PASS computedStyle(prepareNodeInShadowRoot(document.getElementById("non-contenteditable-host1"), null)) is "read-only"
PASS computedStyle(prepareNodeInShadowRoot(document.getElementById("non-contenteditable-host2"), "false")) is "read-only"
PASS computedStyle(prepareNodeInShadowRoot(document.getElementById("non-contenteditable-host3"), "true")) is "read-write"
PASS getUserModifyProperty("child-a") is "read-write"
PASS getUserModifyProperty("child-b") is "read-write"
PASS getUserModifyProperty("child-c") is "read-only"
PASS getUserModifyProperty("child-d") is "read-only"
PASS successfullyParsed is true

TEST COMPLETE
